Moved and renamed mail constraints
