+++ src/TestSuite/Constraint/Email/MailConstraintBase.php
@@ -12,7 +12,7 @@
  * @since         3.7.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\TestSuite\Constraint;
+namespace Cake\TestSuite\Constraint\Email;
 
 use Cake\TestSuite\TestEmailTransport;
 use PHPUnit\Framework\Constraint\Constraint;
