+++ src/TestSuite/Constraint/Email/MailCount.php
@@ -12,12 +12,12 @@
  * @since         3.7.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\TestSuite\Constraint;
+namespace Cake\TestSuite\Constraint\Email;
 
 /**
- * MailCountConstraint
+ * MailCount
  */
-class MailCountConstraint extends MailConstraintBase
+class MailCount extends MailConstraintBase
 {
 
     /**
