+++ src/TestSuite/Constraint/Email/MailSentWith.php
@@ -12,12 +12,12 @@
  * @since         3.7.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\TestSuite\Constraint;
+namespace Cake\TestSuite\Constraint\Email;
 
 /**
- * MailSentWithConstraint
+ * MailSentWith
  */
-class MailSentWithConstraint extends MailConstraintBase
+class MailSentWith extends MailConstraintBase
 {
     protected $method;
